+++ Modules/Intent.Modules.Modelers.Domain.Repositories/Api/ClassExtensionsModel.cs
@@ -1,0 +1,23 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Intent.Metadata.Models;
+using Intent.Modelers.Domain.Api;
+using Intent.Modules.Common;
+using Intent.RoslynWeaver.Attributes;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.ModuleBuilder.Templates.Api.ApiElementExtensionModel", Version = "1.0")]
+
+namespace Intent.Modelers.Domain.Repositories.Api
+{
+    [IntentManaged(Mode.Fully, Signature = Mode.Fully)]
+    public class ClassExtensionsModel : ClassModel
+    {
+        [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
+        public ClassExtensionsModel(IElement element) : base(element)
+        {
+        }
+
+    }
+}